+++ tests/test_random.py
@@ -1353,6 +1353,33 @@ def test_distLookupTable():
             err_msg='Wrong DistDeviate random number sequence for LookupTable with default '
             'interpolant')
 
+    # Test a case with nearly flat probabilities
+    # x and p arrays with and without a small (epsilon) step
+    dx_eps = np.arange(6)
+    dp1_eps = np.zeros(dx_eps.shape)
+    dp2_eps = np.zeros(dx_eps.shape)
+    eps = np.finfo(dp1_eps[0].dtype).eps
+    dp1_eps[0] = 0.5
+    dp2_eps[0] = 0.5
+    dp1_eps[-1] = 0.5
+    dp2_eps[-2] = eps
+    dp2_eps[-1] = 0.5-eps
+    dLookupTableEps1 = galsim.LookupTable(x=dx_eps, f=dp1_eps, interpolant='linear')
+    dLookupTableEps2 = galsim.LookupTable(x=dx_eps, f=dp2_eps, interpolant='linear')
+    d1 = galsim.DistDeviate(testseed, function=dLookupTableEps1, npoints=len(dx_eps))
+    d2 = galsim.DistDeviate(testseed, function=dLookupTableEps2, npoints=len(dx_eps))
+    # If these were successfully created everything is probably fine, but check they create the same
+    # internal LookupTable
+    np.testing.assert_array_almost_equal(
+            d1._inverseprobabilitytable.getArgs(), d2._inverseprobabilitytable.getArgs(), precision,
+            err_msg='DistDeviate with near-flat probabilities incorrectly created '
+                    'a monotonic version of the CDF')
+    np.testing.assert_array_almost_equal(
+            d1._inverseprobabilitytable.getVals(), d2._inverseprobabilitytable.getVals(), precision,
+            err_msg='DistDeviate with near-flat probabilities incorrectly created '
+                    'a monotonic version of the CDF')
+            
+
     # Test filling an image
     d.seed(testseed)
     testimage = galsim.ImageD(np.zeros((3, 1)))